# api-docs/API.yaml
openapi: 3.0.0
info:
  version: '1.0.0'
  title: 'Online-LMS APIs'
  description: 'List of REST APIs Documentation for Online Learning Management System'
  contact:
    name: Palden Sherpa
    email: sherpalden369@gmail.com
servers: 
  - url: http://localhost:5000
    description: Local Server
    
  - url: https://aiex-olms.herokuapp.com
    description: Heroku Server
paths:
...
  /api/admin/book/add-main-category:
    post:
      tags:
        - Book APIs - Admin
      description: Add main book category
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              type: object
              properties:
                categoryName:
                  type: string
                  example: Science and myths
                  
      responses:
        200:
          description: Response on successful addition of main book category
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "book main category addition successful"
  /api/admin/book/add-sub-category:
    post:
      tags:
        - Book APIs - Admin
      description: Add sub category for book
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              type: object
              properties:
                categoryName:
                  type: string
                  example: Animal
                parentID:
                  type: string
                  example: adfadf7987sdfs98798sdf
                  
      responses:
        200:
          description: Response on successful addition of sub category
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "book sub category addition successful"
  /api/admin/book/rename-category:
    put:
      tags:
        - Book APIs - Admin
      description: Rename a category for book
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              type: object
              properties:
                newName:
                  type: string
                  example: Animal
                categoryID:
                  type: string
                  example: adfadf7987sdfs98798sdf
                  
      responses:
        200:
          description: Response on successful rename of a category
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Category rename successful"
  /api/admin/book/all-categories:
    get:
      tags:
        - Book APIs - Admin
      description: Get all category list of Books
      security:
        - bearerAuth: []
      responses:
        200:
          description: Response on successful retrieval of all category list
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "All category list retrieval successful"
  /api/admin/book:
    post:
      tags:
        - Books APIs - Admin
      description: Create a book post
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content: 
          multipart/form-data:
            schema:
              type: object
              properties:
                thumbnail:
                  type: string
                  format: binary
                bookPdf:
                  type: string
                  format: binary
                title:
                  type: string
                  example: Earthquake Resistent Building Design
                authors:
                  type: array
                  items:
                    type: object
                    properties:
                      firstName:
                        type: string
                        example: Palden
                      lastName:
                        type: string
                        example: Sherpa
                edition:
                  type: string
                  example: 2nd(Updated)
                abstract:
                  type: string
                  example: It is a structural design of building.
                categoryID:
                  type: string
                  example: 5454sdfasd54545sdf
                publisher:
                  type: string
                  example: Cambridge University Press
                publicationDate:
                  type: string
                  example: 2020-01-02
                isbn:
                  type: string
                  example: 978-3-16-148410-0
                doi:
                  type: string
                  example: 10.1371/journal.pgen.1001111
      responses:
        200:
          description: Response on successful creation of book post.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: book Post Creation Successful
  /api/admin/book/update-details/{bookID}:
    put:
      tags:
        - Books APIs - Admin
      description: Update a book post
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: bookID
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: Earthquake Resistent Building Design
                authors:
                  type: array
                  items:
                    type: object
                    properties:
                      firstName:
                        type: string
                        example: Palden
                      lastName:
                        type: string
                        example: Sherpa
                edition:
                  type: string
                  example: 2nd(Updated)
                abstract:
                  type: string
                  example: It is a structural design of building.
                categoryID:
                  type: string
                  example: 5454sdfasd54545sdf
                publisher:
                  type: string
                  example: Cambridge University Press
                publicationDate:
                  type: string
                  example: 2020-01-02
                isbn:
                  type: string
                  example: 978-3-16-148410-0
                doi:
                  type: string
                  example: 10.1371/journal.pgen.1001111
      responses:
        200:
          description: Response on successful update of book post.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: book Post Update Successful
  /api/admin/book/update-files/{bookID}:
    put:
      tags:
        - Books APIs - Admin
      description: Update a book files
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: bookID
      requestBody:
        required: true
        content: 
          multipart/form-data:
            schema:
              type: object
              properties:
                thumbnail:
                  type: string
                  format: binary
                pdf:
                  type: string
                  format: binary
                
      responses:
        200:
          description: Response on successful update of book files.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: book files Update Successful
  /api/admin/book/{bookID}:
    get:
      tags:
        - Books APIs - Admin
      description: Get details of a book
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: bookID
      responses:
        200:
          description: Response on successful retrieval of book details.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: book details retrieval Successful

    delete:
      tags:
        - Books APIs - Admin
      description: Delete a book post
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: bookID
      responses:
        200:
          description: Response on successful deletion of book post.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: book Post Delete Successful
  /api/admin/book/all-books/{categoryID}:
    get:
      tags:
        - Books APIs - Admin
      description: Get all books of particular topic
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: categoryID
        - in: query
          name: skips
        - in: query
          name: limit
      responses:
        200:
          description: Response on successful retrieval of books of particular categoryID.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Books of particular categoryID retrieval successful
  /api/admin/book/search:
    get:
      tags:
        - Books APIs - Admin
      description: Search a book post
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: searchKey
        - in: query
          name: skips
        - in: query
          name: limit
     
      responses:
        200:
          description: Response on successful Search of book post.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: book Post Search Successful


#******
#******
#******
#******
#******
#******


#******
#******
#******
#******
#******
#******
...
components:
  securitySchemes:
    bearerAuth:           
      type: http
      scheme: bearer
      bearerFormat: JWT  